Cacheline optimization
Watched your video, figured you'd like to see what if any difference cachelines make. The intuition here is that given the buffers in use we want to guarantee that the memory accesses to the buffers would more than likely belong to (mostly full) cachelines. Ideally we want it such that the access to one region of the buffer doesn't evict the other.

Here I'm making the naive assumption that while multiplying the buffers are iterating from the 0th index forward, therefore the 0th indexs should belong to different cachelines, which should garuntee that the next 8-16 digits will not evict each others access.

I did not test this code or verify its correctness. Hopefully it works first go and makes some improvement (or at least visible difference).